+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 **Version 1.1**
 
-- Added support for INTEGER columns in lookup data.
+- Added support for INTEGER columns in lookup tables.
+- Added support for provided user-defined dictionaries in the configuration.
 
 **Version 1.0**